refactor: match 함수 제거, 상수를 통한 카테고리 이미지 지정 방식으로 변경
